pyproject.toml
[project]
name = "review-classifier"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "kagglehub>=0.3.13",
    "keras>=3.12.0",
    "numpy>=2.3.5",
    "pandas>=2.3.3",
    "scikit-learn>=1.7.2",
    "tensorflow-cpu>=2.20.0",
]
